# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amrutam Telemedicine API",
    "version": "1.0.0",
    "description": "Backend for consultations, doctor availability and prescriptions.\n\n## Idempotency\nEvery non-idempotent write requires an `Idempotency-Key` header. Retrying with the\nsame key replays the original response rather than repeating the side effect. Keys\nare scoped per user and expire after 24 hours.\n\n## Errors\nEvery error is `{ \"error\": { \"code\", \"message\", \"requestId\" } }`. Quote `requestId`\nin a support request — it correlates the response with logs and traces.\n\n## Rate limits\nResponses carry `RateLimit-Limit` and `RateLimit-Remaining`. Exceeding a limit returns\n429 with `Retry-After`.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /auth/login. Expires in 15 minutes; rotate with /auth/refresh."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "SLOT_UNAVAILABLE"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "string",
            "example": "15m"
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "tokenType"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a patient or doctor account",
        "description": "Admin accounts cannot be self-registered. A doctor account is created unverified and cannot accept bookings until an administrator verifies the medical registration number.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 128
                  },
                  "fullName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "patient",
                      "doctor"
                    ],
                    "default": "patient"
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^\\+?[0-9]{8,15}$"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "fullName"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "role"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — password policy or malformed input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange credentials for a token pair",
        "description": "Returns 401 both for an unknown account and a wrong password, deliberately, so the endpoint cannot be used to enumerate accounts. Five failures lock the account for 15 minutes. If MFA is enabled, `totp` is required.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "totp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TokenPair"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials, or MFA code required/incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Account locked or not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Rotate a refresh token",
        "description": "Single-use. Presenting an already-used token is treated as theft: the entire token family is revoked and every session for that user ends.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 512
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TokenPair"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, revoked, or reused token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke a refresh-token family",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 512
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Logged out (idempotent)"
          }
        }
      }
    },
    "/api/v1/auth/mfa/setup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Begin TOTP enrolment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Returns a secret and an otpauth:// URI for a QR code. MFA is not active until confirmed.",
        "responses": {
          "200": {
            "description": "Enrolment started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "secret": {
                          "type": "string"
                        },
                        "otpauthUri": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "secret",
                        "otpauthUri"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "MFA already enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/confirm": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Activate TOTP after verifying a code",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "totp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "totp"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA enabled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "mfaEnabled": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "mfaEnabled"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Current token identity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Identity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "mfaSatisfied": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "role",
                        "mfaSatisfied"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}